+++ ragnar_drivers/package.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <package>
-  <name>ros_ragnar</name>
+  <name>ragnar_drivers</name>
   <version>0.0.0</version>
-  <description>The ros_ragnar package</description>
+  <description>The ragnar_drivers package</description>
 
   <!-- One maintainer tag required, multiple allowed, one person per tag --> 
   <!-- Example:  -->
@@ -19,7 +19,7 @@
   <!-- Url tags are optional, but mutiple are allowed, one per tag -->
   <!-- Optional attribute type can be: website, bugtracker, or repository -->
   <!-- Example: -->
-  <!-- <url type="website">http://wiki.ros.org/ros_ragnar</url> -->
+  <!-- <url type="website">http://wiki.ros.org/ragnar_drivers</url> -->
 
 
   <!-- Author tags are optional, mutiple are allowed, one per tag -->
